Extracts more methods
